# webhook/api/swagger-codegen/lettuce-webhook-openapi.yaml
openapi: 3.0.3
info:
  title: Plants Soil Lettuce (Webhook Engine)  - OpenAPI 3.0
  description: |-
    This is a Webhook Engine Core API based on the OpenAPI 3.0 specification.
    You can register as publisher and trigger (publish) events (message carried) as needed, all subscribers could receive the events instantly.
    You can register as subscriber with webhooks to listen on and receive events (messages carried) from publishers you subscribed.

    Some useful links:
    - [The Lettuce Webhook Engine repository](https://github.com/plants-soil/lettuce-webhook-hub)
  contact:
    email: danialdy@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://lettuce.plantssoil.com/api/v1
...
paths:
...
  /engine/trigger:
    post:
      tags:
        - engine
      summary: Trigger a webhook event and post a message
      description: Trigger a webhook event and post a message
      operationId: trigger
      requestBody:
        description: Trigger a webhook event and post a message
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookMessage'
          application/xml:
            schema:
              $ref: '#/components/schemas/WebhookMessage'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WebhookMessage'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'          
            application/xml:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation exception
      security:
        - api_key: []
        - engine_auth:
            - write:engine
            - read:engine
  /engine/allPublisherWebhookLogs:
    get:
      tags:
        - engine
      summary: Find all webhook logs which triggered by specific publisher with pagination
      description: Will find the webhook logs on the page specified (page, pageSize)
      operationId: findAllPublisherWebhookLogs
      parameters:
        - name: publisherId
          in: query
          description: ID of publisher which triggered webhooks
          required: true
          schema:
            type: string
        - name: dataGroup
          in: query
          description: Data group (name) which triggered webhooks
          schema:
            type: string
        - name: page
          in: query
          description: The page number of webhook logs query result
          required: true
          explode: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          description: Maximum webhook logs on current page
          required: true
          explode: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookLog'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookLog'
        '400':
          description: Validation exception
      security:
        - webhooklog_auth:
            - write:webhooklogs
            - read:webhooklogs
  /engine/allSubscriberWebhookLogs:
    get:
      tags:
        - engine
      summary: Find all webhook logs which dispatched to specific subscriber with pagination
      description: Will find the webhook logs on the page specified (page, pageSize)
      operationId: findAllSubscriberWebhookLogs
      parameters:
        - name: webhookId
          in: query
          description: ID of webhook which receive events & messages
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: The page number of webhook logs query result
          required: true
          explode: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          description: Maximum webhook logs on current page
          required: true
          explode: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookLog'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookLog'
        '400':
          description: Validation exception
      security:
        - webhooklog_auth:
            - write:webhooklogs
            - read:webhooklogs
  /engine/webhookLines:
    get:
      tags:
        - engine
      summary: Find all webhook log lines which belong to specific request & webhook
      description: Will find the webhook log lines
      operationId: findWebhookLogLines
      parameters:
        - name: requestId
          in: query
          description: ID of request
          required: true
          schema:
            type: string
        - name: webhookId
          in: query
          description: ID of webhook
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookLogLine'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookLogLine'
        '400':
          description: Validation exception
      security:
        - webhooklog_auth:
            - write:webhooklogs
            - read:webhooklogs
components:
  schemas:
...
    WebhookLog:
      description: The webhook logs, all events (which carries messages) will be logged when publisher trigger events
      type: object
      properties:
        eventType:
          type: string
          description: The event type publisher triggered
          example: order.created, or order.paid, order.delivered, product.created, etc.
        eventTag:
          type: string
          description: The event tag publisher triggered
          example: product, order, etc.
        contentType:
          type: string
          description: The content type publisher triggered
          example: application/json, application/xml, etc.
        charset:
          type: string
          description: The message charset publisher triggered
          example: utf-8, etc.
        requestId:
          type: string
          description: ID of request
          example: 6791d42ed3d172552ed09459, maximum 24 chars
        payload:
          type: string
          description: The payload (request body) publisher triggered
          example: {"orderId": "1", "orderTime": "2025-02-08T18:18:18", "customerId": "6791d42ed3d172552ed09459", "totalAmount": 120, "orderLines": [{"rowno": 1, "sku": "100001", "quantity": 1, "unitPrice": 40}, {"rowno": 1, "sku": "100002", "quantity": 1, "unitPrice": 80}]}
        webhookStatus:
          type: string
          description: The webhook status
          example: triggered
        triggerTime:
          type: string
          description: The time when publisher triggered the event
          format: date-time
          example: 2025-01-23T14:56:07Z
      xml:
        name: webhookLog
    WebhookLogLine:
      description: The webhook log line, log line will be added when event(message) is dispatched, got response from webhook, or exception happend
      type: object
      properties:
        logType:
          type: string
          description: The log type
          example: dispatched, success, fail, etc.
        information:
          type: string
        logTime:
          type: string
          description: The log time when the log line is added
          format: date-time
          example: 2025-01-23T14:56:07Z
        tryTime:
          type: integer
          description: How many times the event(with message) is dispatched
          format: int32
          example: 1, 2, 3
      xml:
        name: webhookLogLine
    WebhookMessage:
      description: The webhook message which triggered by publisher
      required:
        - publisherId
        - version
        - eventType
        - requestId
        - payload
      type: object
      properties:
        publisherId:
          type: string
          description: Valid publisher ID (who trigger the event and message), required
          example: 6791d42ed3d172552ed09459, maximum 24 chars
        version:
          type: string
          description: Valid publisher version (who trigger the event and message), required
          example: 1.0.0
        eventType:
          type: string
          description: Valid event type (should already been registered), required
          example: order.created, or order.paid, order.delivered, product.created, etc.
        eventTag:
          type: string
          description: Valid event tag, should already been registered
          example: product, order, etc.
        contentType:
          type: string
          description: The content type (format type of payload)
          example: application/json, application/xml, etc.
        charset:
          type: string
          description: The message charset (encoding of payload)
          example: utf-8, etc.
        dataGroup:
          type: string
          description: Data group (the identity of the data group if the publisher support multiple data group, usally could be business unit ID / merchant ID), should already been registered
          example: 6791d42ed3d172552ed09459
        requestId:
          type: string
          description: ID of request, which used to identify one request, required. Webhook could use this requestId to verify duplicated consumption, this verification should be always needed, because webhook engine will try to call webhook multiple times util got response from webhook.
          example: 6791d42ed3d172552ed09459
        payload:
          type: string
          description: The message (request body) publisher triggered, required
          example: {"orderId": "1", "orderTime": "2025-02-08T18:18:18", "customerId": "6791d42ed3d172552ed09459", "totalAmount": 120, "orderLines": [{"rowno": 1, "sku": "100001", "quantity": 1, "unitPrice": 40}, {"rowno": 1, "sku": "100002", "quantity": 1, "unitPrice": 80}]}
      xml:
        name: webhookMessage
    ApiResponse:
      description: The API response object which hold the response code and messages
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: integer
          format: int32
          description: The response code
          example: 200
        message:
          type: string
          description: The response message
          example: called successfully
        data:
          type: object
          description: The data attached to the response
      xml:
        name: apiResponse
  securitySchemes:
    organization_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:organizations: modify organizations in your account
            read:organizations: read your organizations
    publisher_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:publishers: modify publishers in you account
            read:publishers: read your publishers
    event_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:events: modify events in your account
            read:events: read your events
    datagroup_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:datagroups: modify data groups in your account
            read:datagroups: read your data groups
    subscriber_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:subscribers: modify subscribers in your account
            read:subscribers: read your subscribers
...